requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
h2>=4.1.0
grpcio-tools>=1.78.0
python-dotenv>=1.0.0
protobuf>=4.25.0
pydantic>=2.5.0

# WebSocket support for Orchids
websockets>=12.0

# JWT parsing for Orchids
PyJWT>=2.8.0
